Added skeleton index.html file & script.js
Created a getComputerChoice function which will give us a random number
from 1-3. playerSelection function will assign a number to the players
choice. For example, rock will equal 1, paper will equal 2. This will
be useful for the game where we will compare the choices for a win,
loss, or a draw.